Implement Meka-Tool Shearing module (https://github.com/mekanism/Mekanism-Feature-Requests/issues/65), and MekaSuit Geothermal Generator module. Also refactor the silk touch module to use the system we designed for enchantment based modules for the frost walker boots so as to play nicer with other mods #7243. Also exposed the following abilities to custom modules: - Ability to define entity interactions for the Meka-Tool - Ability to define custom dispenser behaviors for the Meka-Tool and MekaSuit (equipping to armor stands/players still takes precedent for the MekaSuit) - A way to do things when the module's enabled state changes (frost walker module can now be disabled without removing it) - A way to more easily create custom modules that can prevent certain types of damage Also fix our recipe cache invalidation listener potentially running slightly earlier than it was meant to and module setting syncing to server not being properly validated server side. Make MekaTool and Atomic Disassembler mine at a speed of zero when they don't have any energy

+++ src/main/java/mekanism/common/content/gear/mekatool/ModuleExcavationEscalationUnit.java
@@ -49,7 +49,7 @@ public class ModuleExcavationEscalationUnit implements ICustomModule<ModuleExcav
         }
     }
 
-    public double getEfficiency() {
+    public float getEfficiency() {
         return excavationMode.get().getEfficiency();
     }
